UefiPayloadPkg/Library/PciHostBridgeLib: Add warning for coreboot/EDK2 incompability
EDK2 expects PCI bars to be continuously allocated without "gaps" of
reserved memory in between.
coreboot places PCI bars anywhere in the PCI MMIO space, interleaved
with MMCONF and reserved I/O MMIO space.

Warn about this behaviour and refuse to add the BAR to the PCI aperture
as it would cause the PciHostBridgeDxe to fail.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>