[WIP] OvmfPkg/Csm: Do not assert on failure to allocate 0x0 region
This might already be reserved, for either safety purposes or because it
contains other data. For example, coreboot low tables are here.

Both the old solution, that omitted this clearing, and the new one, that
simply avoids asserting, have been tested on QEMU. It's likely that
further testing wouldn't be needeed, at least conceptually, since
SeaBIOS will initialise the memory either way.

Previously, this was untested on physical hardware, and I was unclear if
other regions are reserved too (concerned about A-F segments). This may
be fine now.

Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>