[WIP] UefiPayloadPkg: Add CSM support
Add initial support for loading a CSM, using modules in OvmfPkg.

Tested on QEMU; SeaBIOS can boot a live CD. Note that some features are
broken (like PS/2 mice). Details about the protocols involved in this
follow, and it's hoped that by finalising them, everything will work.

The legacy region and legacy interrupt protocol support are refinements
of my own work in 2020/2021, which was partially derived and/or inspired
by implementations of some functions in OpenCore, now with AMD support
basically copied from Dasharo.

More work is needed on the BBS handover tables. Entries are always
considered as hard disks, and if multiple are present, attempting to
boot any will always boot the first found. Perhaps this is implied by
the CSM spec, but then we really need to refine the BBS tables so that
we can consider all kinds of disk types.

TODO: Don't hardcode ITSS details, get them from coreboot.

Further testing must be done on physical hardware.

Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>