[WIP] MdeModulePkg: Add new SerialPortLibMem
SerialPortLibMem is an implementation of a SerialPortLib designed
for obtaining debug messages in commercial systems or where debug
facilities might otherwise be restricted. It writes to a memory buffer
and writes the address and offsets to variables at ExitBootServices()
for the user's convenience.

The PEI instance initially creates a small buffer, due to temporary
memory size restrictions, but may create and move to a larger buffer at
PeiMemoryDiscovered. The DXE instance will copy the HOB to its own
allocation to eventually provide to the user.

No SMM instance has been written yet. Further consideration is required
for this phase.

Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>