MdeModulePkg: PciBusDxe: Drop `RemoveRejectedPciDevices()`
There is an assumption in `IsPciDeviceRejected()` (at least for I/O)
that a BAR with all read-write bits set is invalid. However, this is not
the case: when a resource is size-aligned down from the top, this may
also be the case.

This code causes the rejection of the iGPU (and therefore, display) in
some cases with the UefiPayloadPkg, when the coreboot
ESOURCE_ALLOCATION_TOP_DOWN config is enabled. It remains to be seen
whether this code addresses some other issue.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>