UefiPayloadPkg/BlParseLib: Add parsing of firmware info
The implementation is for CbParseLib, SblParseLib returns an error.

coreboot's CB_TAG_FW_INFO is a machine-friendly version of a system
firmware component.  A component is identified by a GUID.  This is meant
to be primarily a source of information for ESRT.

The following coreboot options translate directly into fields of the
structure with information:
 - CONFIG_DRIVERS_EFI_MAIN_FW_GUID
 - CONFIG_DRIVERS_EFI_MAIN_FW_VERSION
 - CONFIG_DRIVERS_EFI_MAIN_FW_LSV
 - CONFIG_ROM_SIZE

The first 3 options depend on CONFIG_DRIVERS_EFI_FW_INFO.

Version as a string depends on CONFIG_LOCALVERSION as well as version of
the code base (misnamed as KERNELVERSION in coreboot's build system).

Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>