UefiPayloadPkg: CustomFdtNodeParserLib: FitIsHobNeed() as allowlist
This function is easier to read as a filtered allowlist. But more than that,
this allows us to be explicit about what is really needed.
Previously, some HOB types slipped through.

One HobType that was accidentally let through is EFI_HOB_TYPE_FV. On its own,
this is already a bug, since it's not the proper way for Platform Init
to provide FVs to UPL anyways: the spec outlines
`/options/upl-image/image@<addr>` as the way for passing FVs across
(as opposed to `/options/upl-image@<addr>` which points at whole-FIT binaries).

Later, I added code to perform that task of finding FVs in the UPL FDT.
I made them mutually exclusive, since if Platform Init extracted some FVs,
then the FIT header no longer describes them (the offset, `load`,
`entry` values, etc will be incorrect). Therefore, it makes sense to
only read images from the FIT header if FV HOBs weren't already created
from the UPL FDT.

However, between these two points, we encounter a bug on EDK2-based
Platform Init that uses `/options/upl-image@<addr>`: there already are FV HOBs
in the system (ones from Platform Init that shouldn't be there),
so we don't parse the FIT header at all. Which means that the DXE core
is never discovered, which is a fatal bug.

The best solution is to start being explicit about which HOBs are needed
and should be copied through for UPL.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>