BaseTools/GenerateCapsule.py: skip pointless verification on --decode
When a capsule is signed, --decode printed a warning if neither openssl
nor signtool was configured via command-line options.  And yet, it then
proceeded to attempt the verification anyway.

Current implementation only ever uses --trusted-public-cert during
verification, so there could be some logic to this behaviour.  However,
it's still confusing, so don't attempt to verify anything if it's not
expected to work.

Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>