UefiPayloadPkg: Drop AARCH64's resolution PCDs overrides
There is no architectural need to override these PCDs, and platforms can
override it on the `build` command line if needed. AARCH64 now joins x86
in using the entire screen.

The setup menu is now 100 x 31. coreboot optionally overrides this
using the command line, and if upstream wants this, we can do it in a
follow-up.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>